GSoft user is unable to promote Slack connection to use real-time https://trello.com/c/8A7JiKPY

+++ vs/src/CodeStream.VisualStudio/Services/AuthenticationService.cs
@@ -1,8 +1,11 @@
 ﻿using CodeStream.VisualStudio.Core.Events;
 using CodeStream.VisualStudio.Core.Extensions;
+using CodeStream.VisualStudio.Core.LanguageServer;
 using CodeStream.VisualStudio.Core.Logging;
 using CodeStream.VisualStudio.Core.Models;
 using CodeStream.VisualStudio.Core.Services;
+using Microsoft;
+using Microsoft.VisualStudio.ComponentModelHost;
 using Serilog;
 using System;
 using System.ComponentModel.Composition;
@@ -13,6 +16,13 @@ namespace CodeStream.VisualStudio.Services {
 	[PartCreationPolicy(CreationPolicy.Shared)]
 	public class AuthenticationService : IAuthenticationService {
 		private static readonly ILogger Log = LogManager.ForContext<AuthenticationService>();
+		private readonly IServiceProvider _serviceProvider;
+
+		[ImportingConstructor]
+		public AuthenticationService(
+				[Import(typeof(Microsoft.VisualStudio.Shell.SVsServiceProvider))] IServiceProvider serviceProvider) {
+			_serviceProvider = serviceProvider;
+		}
 
 		[Import]
 		public ISessionService SessionService { get; set; }
@@ -29,29 +39,40 @@ namespace CodeStream.VisualStudio.Services {
 		[Import]
 		public IWebviewUserSettingsService WebviewUserSettingsService { get; set; }
 
-		public async System.Threading.Tasks.Task LogoutAsync() {
+		public async System.Threading.Tasks.Task LogoutAsync(SessionSignedOutReason reason = SessionSignedOutReason.UserSignedOutFromWebview) {
 			Log.Information($"{nameof(LogoutAsync)} starting");
 			try {
 				try {
 					SessionService.SetState(SessionState.UserSigningOut);
 				}
-				catch(Exception ex) {
+				catch (Exception ex) {
 					Log.Warning(ex, $"{nameof(LogoutAsync)} - SetState");
 				}
 
 				try {
 					EventAggregator.Publish(new SessionDidStartSignOutEvent());
 				}
-				catch(Exception ex) {
+				catch (Exception ex) {
 					Log.Warning(ex, $"{nameof(LogoutAsync)} - {nameof(SessionDidStartSignOutEvent)}");
 				}
-				
-				try {
-					var settingsService = SettingsServiceFactory.Create();
-					await CredentialsService.DeleteAsync(settingsService.ServerUrl.ToUri(), settingsService.Email);
-				}
-				catch (Exception ex) {
-					Log.Warning(ex, $"{nameof(LogoutAsync)} - credentials");
+
+				if (
+					reason == SessionSignedOutReason.UserSignedOutFromWebview ||
+					reason == SessionSignedOutReason.UserSignedOutFromExtension ||
+					reason == SessionSignedOutReason.ReAuthenticating) {
+					try {
+						var settingsService = SettingsServiceFactory.Create();
+						await CredentialsService.DeleteAsync(settingsService.ServerUrl.ToUri(), settingsService.Email);
+					}
+					catch (Exception ex) {
+						Log.Warning(ex, $"{nameof(LogoutAsync)} - credentials");
+					}
+					try {
+						WebviewUserSettingsService.DeleteTeamId(SessionService.SolutionName);
+					}
+					catch (Exception ex) {
+						Log.Error(ex, "could not delete teamId");
+					}
 				}
 
 				try {
@@ -61,21 +82,20 @@ namespace CodeStream.VisualStudio.Services {
 					Log.Error(ex, $"{nameof(LogoutAsync)} - agent");
 				}
 
-				try {
-					WebviewUserSettingsService.DeleteTeamId(SessionService.SolutionName);					
-				}
-				catch(Exception ex) {
-					Log.Error(ex, "could not delete teamId");
-				}
+				if (reason == SessionSignedOutReason.UserSignedOutFromWebview ||
+					reason == SessionSignedOutReason.UserSignedOutFromExtension
+				) {
+					//don't call this when ReAuthenticationg -- don't want to show the login screen
+					try {
 
-				try {
 #pragma warning disable VSTHRD103 // Call async methods when in an async method
-					// it's possible that this Logout method is called before the webview is ready -- enqueue it
-					WebviewIpc.EnqueueNotification(new HostDidLogoutNotificationType());
+						// it's possible that this Logout method is called before the webview is ready -- enqueue it
+						WebviewIpc.EnqueueNotification(new HostDidLogoutNotificationType());
 #pragma warning restore VSTHRD103 // Call async methods when in an async method
-				}
-				catch (Exception ex) {
-					Log.Error(ex, $"{nameof(LogoutAsync)} - {nameof(HostDidLogoutNotificationType)}");
+					}
+					catch (Exception ex) {
+						Log.Error(ex, $"{nameof(LogoutAsync)} - {nameof(HostDidLogoutNotificationType)}");
+					}
 				}
 
 				try {
@@ -92,6 +112,19 @@ namespace CodeStream.VisualStudio.Services {
 					Log.Error(ex, $"{nameof(LogoutAsync)} - {nameof(SessionLogoutEvent)}");
 				}
 
+				if (reason == SessionSignedOutReason.UserSignedOutFromWebview ||
+					reason == SessionSignedOutReason.UserSignedOutFromExtension ||
+					reason == SessionSignedOutReason.ReAuthenticating) {
+					var componentModel = _serviceProvider.GetService(typeof(SComponentModel)) as IComponentModel;
+					Assumes.Present(componentModel);
+					var languageServerClientManager = componentModel.GetService<ILanguageServerClientManager>();
+					if (languageServerClientManager != null) {
+						await languageServerClientManager.RestartAsync();
+					}
+					else {
+						Log.IsNull(nameof(ILanguageServerClientManager));
+					}
+				}
 				Log.Information($"{nameof(LogoutAsync)} completed");
 			}
 			catch (Exception ex) {
